calculateRect - helper
